+++ pages/oauth/signup/[provider].tsx
@@ -1,5 +1,8 @@
 import OauthSignUpForm from "@/src/components/auth/oauth/OauthSignUpForm";
+import useCheckSignIn from "@/src/hooks/useCheckSignIn";
 
 export default function OauthSignUp() {
-  return <OauthSignUpForm />;
+  const isLoggedIn = useCheckSignIn();
+
+  return !isLoggedIn && <OauthSignUpForm />;
 }
