+++ pages/signup/index.tsx
@@ -1,18 +1,8 @@
 import SignUpForm from "@/src/components/auth/signUp/SignUpForm";
-import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
+import useCheckSignIn from "@/src/hooks/useCheckSignIn";
 
 export default function SingUp() {
-  const router = useRouter();
-  const [isLoggedIn, setIsLoggedIn] = useState(true);
-  useEffect(() => {
-    if (localStorage.getItem("accessToken")) {
-      router.push("/");
-      return;
-    } else {
-      setIsLoggedIn(false);
-    }
-  }, []);
+  const isLoggedIn = useCheckSignIn();
 
   return !isLoggedIn && <SignUpForm />;
 }
